+++ src/app/locales/en/components/modals/apply-atm-record-dump-modal/dump-details.js
@@ -1,0 +1,14 @@
+export default {
+  header: {
+    atmLambda: 'Lambda details',
+    atmWorkflowSchema: 'Workflow details',
+  },
+  name: 'Name',
+  summary: 'Summary',
+  revisionNumber: 'Revision',
+  unknownValue: 'none',
+  error: {
+    atmLambda: 'Uploaded file is not a valid lambda dump.',
+    atmWorkflowSchema: 'Uploaded file is not a valid workflow dump.',
+  },
+};