+++ src/tests/helpers/workflows/generate-atm-lambda-dump.js
@@ -1,0 +1,15 @@
+export default function generateAtmLambdaDump() {
+  return {
+    schemaFormatVersion: 2,
+    revision: {
+      schemaFormatVersion: 2,
+      atmLambdaRevision: {
+        name: 'l1',
+        summary: 'summary',
+        state: 'stable',
+      },
+      originalRevisionNumber: 3,
+    },
+    originalAtmLambdaId: 'l1id',
+  };
+}